requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
fastapi
uvicorn
python-multipart
pillow
requests
